Updated/Enhanced Translation and Research System
Huge change, researches are now chosen in translation table, then you
can translate them and get them as items. Once you have them translated
you can open the page and read the research or add it inside of your h.
book. Once you added it, you should be able to read it. Be aware that
there are things that are not 100% complete or/and could be done better.